adding test for dockerfile
