--- internal/genny/newapp/core/errors.go
@@ -1,6 +1,0 @@
-package core
-
-import "fmt"
-
-// ErrNotInGoPath is thrown when not using go modules outside of GOPATH
-var ErrNotInGoPath = fmt.Errorf("currently not in a $GOPATH")
